requirements.txt
zpython-dotenv~=0.19.1
SQLAlchemy~=1.4.25
python-dotenv~=0.19.1
vk_api
time
re
datetime